move clock/alarm modules separate files
